requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
dataclasses==0.6
Django==5.1.1
django-cors-headers==4.4.0
django-oauth-toolkit==3.0.1
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.15.2
dnspython==2.6.1
idna==3.10
jwcrypto==1.5.6
mongoengine==0.29.1
oauthlib==3.2.2
pycparser==2.22
pymongo==4.9.1
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3

